module-cero/pyproject.toml
[project]
name = "module-cero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=2025.4.1",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.50",
    "langchain-groq>=0.3.2",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "tavily-python>=0.5.4",
]
